CLEANUP: Move flash drivers to their own directory
