+++ mk/source.mk
@@ -226,6 +226,7 @@ endif
 endif
 
 RX_SRC = \
+            $(addprefix drivers/rx/,$(notdir $(wildcard $(SRC_DIR)/drivers/rx/*.c))) \
             rx/cc2500_common.c \
             rx/cc2500_frsky_shared.c \
             rx/cc2500_frsky_d.c \
@@ -234,22 +235,12 @@ RX_SRC = \
             rx/cc2500_redpine.c \
             rx/a7105_flysky.c \
             rx/cyrf6936_spektrum.c \
-            drivers/rx/expresslrs_driver.c \
             rx/expresslrs.c \
             rx/expresslrs_common.c \
-            rx/expresslrs_telemetry.c \
-            drivers/rx/rx_cc2500.c \
-            drivers/rx/rx_a7105.c \
-            drivers/rx/rx_cyrf6936.c \
-            drivers/rx/rx_sx127x.c \
-            drivers/rx/rx_sx1280.c
+            rx/expresslrs_telemetry.c
 
 FLASH_SRC += \
-            drivers/flash.c \
-            drivers/flash_m25p16.c \
-            drivers/flash_w25n.c \
-            drivers/flash_w25q128fv.c \
-            drivers/flash_w25m.c \
+            $(addprefix drivers/flash/,$(notdir $(wildcard $(SRC_DIR)/drivers/flash/*.c))) \
             io/flashfs.c
 
 SDCARD_SRC += \
