+++ src/main/drivers/flash/flash_w25m.c
@@ -36,8 +36,8 @@
 
 #include "common/maths.h"
 #include "drivers/bus_spi.h"
-#include "drivers/flash.h"
-#include "drivers/flash_impl.h"
+#include "flash.h"
+#include "flash_impl.h"
 #include "drivers/io.h"
 #include "drivers/time.h"
