+++ src/main/fc/init.c
@@ -56,7 +56,7 @@
 #include "drivers/dma.h"
 #include "drivers/dshot.h"
 #include "drivers/exti.h"
-#include "drivers/flash.h"
+#include "drivers/flash/flash.h"
 #include "drivers/inverter.h"
 #include "drivers/io.h"
 #include "drivers/light_led.h"
